eval: use a function-global variable for location information
Location information was previously only extracted from the closure of
the application cases, and thus local to the code block handling
function application. At this point, the original Scheme_App*_Rec object
has been traversed, so its own location information is lost.

We move the location information to a global variable so that we can
communicate it across jumps in the eval function: in the future we
will set it when the App_Rec object is still available, then use it
after the jump to the application evaluation case.